range selector added to settings, ui improved

+++ app/lib/config/Config.dart
@@ -4,6 +4,6 @@ class Config {
   static int SENSOR_PER_PAGE_COUNT = 12;
   static const int SENSOR_PER_ROW_COUNT = 3;
   // Limits can be changed in settings screen
-  static int UPPER_LIMIT = 20;
-  static int LOWER_LIMIT = 0;
+  static double UPPER_LIMIT = 20;
+  static double LOWER_LIMIT = 0;
 }
